More cleanup and testing
